go/merkle/encoder.go
package merkle

import (
	"crypto/hmac"
	cryptorand "crypto/rand"
	"crypto/sha256"
	"crypto/sha512"

	"github.com/pkg/errors"
)

type Encoder struct {
	encodingType EncodingType
}

func NewEncoder(encodingType EncodingType) *Encoder {
	return &Encoder{encodingType: encodingType}
}

func (e *Encoder) BlindedPreimage(leaf Leaf, key Key, secret Secret) (BlindedPreimage, error) {
	switch e.encodingType {
	case EncodingTypeBlindedSHA256v1:
		h := hmac.New(sha256.New, secret.Secret)
		h.Write(key.Key)
		return NewBlindedPreimage(leaf, h.Sum(nil))
	case EncodingTypeBlindedSHA512_256v1:
		h := hmac.New(sha512.New512_256, secret.Secret)
		h.Write(key.Key)
		return NewBlindedPreimage(leaf, h.Sum(nil))
	default:
		return BlindedPreimage{}, errors.Errorf("unknown encoding type %q", e.encodingType)
	}
}

func (e *Encoder) Hash(preimage BlindedPreimage) ([]byte, error) {
	b, err := preimage.LeafContainer.Serialize()
	if err != nil {
		return nil, err
	}
	switch e.encodingType {
	case EncodingTypeBlindedSHA256v1:
		h := hmac.New(sha256.New, preimage.BlindedEntropy)
		h.Write(b)
		return h.Sum(nil), nil
	case EncodingTypeBlindedSHA512_256v1:
		h := hmac.New(sha512.New512_256, preimage.BlindedEntropy)
		h.Write(b)
		return h.Sum(nil), nil
	default:
		return nil, errors.Errorf("unknown encoding type %q", e.encodingType)
	}
}

func (e *Encoder) Encode(leaf Leaf, key Key, secret Secret) ([]byte, error) {
	preimage, err := e.BlindedPreimage(leaf, key, secret)
	if err != nil {
		return nil, err
	}
	return e.Hash(preimage)
}

func (e *Encoder) GenerateSecret() (Secret, error) {
	switch e.encodingType {
	case EncodingTypeBlindedSHA256v1, EncodingTypeBlindedSHA512_256v1:
		secret := make([]byte, 32)
		_, err := cryptorand.Read(secret)
		return NewSecret(secret), err
	default:
		return Secret{}, errors.Errorf("unknown encoding type %q", e.encodingType)
	}
}
